extracted submit logic outside upload
