+++ src/web/src/Pages/PostPage.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect, useRef, useContext } from 'react';
 import { useLocation } from 'react-router-dom';
 import { Header, Footer, SideNavWrapper } from '../Components';
 import { CDN_URL } from '../Constants';
-import { PostType } from '../db.types';
+import { PostType } from '../global.types';
 import { AuthContext } from '../App';
 import axios from 'axios';
 import gear from '../assets/svg/gear.svg';
