+++ src/web/src/Pages/ProfilePage.tsx
@@ -11,12 +11,11 @@ import {
 import { getUserByUsername } from '../util/getUser';
 import { GAMES } from '../Constants';
 import { AuthContext } from '../App';
-import { UserType, PostType } from '../db.types';
+import { UserType, PostType } from '../global.types';
 import { sendVerificationEmail } from '../util/sendVerificationEmail';
 import { CiEdit } from 'react-icons/ci';
 import axios from 'axios';
 import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 
 const ProfilePage = () => {
    const { id } = useParams<{ id: string }>();
