+++ src/web/src/Pages/auth/VerifyEmail.tsx
@@ -3,7 +3,6 @@ import axios from 'axios';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { Header, SideNavWrapper } from '../../Components';
 import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 
 const VerifyEmail: React.FC = () => {
    const [verificationCode, setVerificationCode] = useState<string>('');
