+++ src/web/src/Pages/game/Valorant.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import Posts from '../../Components/Posts';
-import { PostType } from '../../db.types';
+import { PostType } from '../../global.types';
 import axios from 'axios';
 import {
    Header,
@@ -37,10 +37,7 @@ const Valorant: React.FC = () => {
    return (
       <>
          {value && open && (
-            <ValorantPopup
-               setOpen={setOpen}
-               setValue={setValue}
-            />
+            <ValorantPopup setOpen={setOpen} setValue={setValue} />
          )}
          <div className="flex flex-col min-h-screen">
             <Header />
