+++ src/web/src/global.types.ts
